Разнес константы по компонентам
