+++ src/consts/index.js
@@ -1,12 +1,4 @@
-const DELAY_ON_MOUSE_ENTER = 2000;
-
-const TabNames = {
-  OVERVIEW: `Overview`,
-  DETAILS: `Details`,
-  REVIEWS: `Reviews`,
-};
-
-const FilterNames = {
+const FilterName = {
   ALL: `All genres`,
 };
 
@@ -14,32 +6,8 @@ const KeyName = {
   ESC: `Escape`,
 };
 
-const FilmRating = {
-  BAD: `Bad`,
-  NORMAL: `Normal`,
-  GOOD: `Good`,
-  VERY_GOOD: `Very good`,
-  AWESOME: `Awesome`,
-};
-
-const ErrorMessage = {
-  LOGIN: `Введите корректный email`,
-  PASSWORD: `Введите корректный пароль`,
-};
-
-const FILTER_ALL_GENRES_NAME = `All genres`;
-const SHOW_FILM_CARD_BY_START = 8;
-const SHOW_FILM_CARD_BY_BUTTON = 8;
-
 
 export {
-  DELAY_ON_MOUSE_ENTER,
-  TabNames,
-  FilterNames,
   KeyName,
-  FilmRating,
-  ErrorMessage,
-  FILTER_ALL_GENRES_NAME,
-  SHOW_FILM_CARD_BY_START,
-  SHOW_FILM_CARD_BY_BUTTON,
+  FilterName,
 };
